semver: reduce memory usage in Parse, Clean, and Compare
Factor the internals of Clean differently so we don't need to re-do work.  This
saves a bunch of memory in the common case.

Add deferred-text error types so we only have to pay the cost of one error box
rather than a new wrapper and a string format.

Add a test for the error-generating path of Parse.

Benchmarks (vs. v0.1.5)

    time/op           before   after   delta
    Parse/Short-10    52.04n   47.42n  -8.88%
    Parse/Long-10     183.8n   184.5n  +0.38%
    Clean/Valid-10    71.41n   70.27n  -1.60%
    Clean/Invalid-10  204.5n   243.7n  +19.17%
    Compare/Eq-10     100.00n  65.41n  -34.59%
    Compare/Neq-10    103.40n  66.91n  -35.29%

    bytes/op          before   after   delta
    Clean/Invalid-10  104.00   56.00   -46.15%
    Compare/Eq-10     96.00    0.00    -100.00%
    Compare/Neq-10    96.00    0.00    -100.00%

    allocs/op         before   after   delta
    Clean/Invalid-10  5.000    4.000   -20.00%
    Compare/Eq-10     2.000    0.000   -100.00%
    Compare/Neq-10    2.000    0.000   -100.00%